# Fix Integer and StdFactKB memory leak with patch from #9294