<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234". See
https://github.com/blog/1506-closing-issues-via-pull-requests . Please also
write a comment on that issue linking back to this pull request once it is
open. -->
Fixes #15778 
Fixes #9292

#### Brief description of what is fixed or changed
The unbounded and adhoc cache infrastructure (_intcache) was replaced with lru_cache in numbers.py.
Inline with this, the test_integer_cache was removed from test_numbers.py   

#### Other comments


#### Release Notes

<!-- Write the release notes for this release below. See
https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more information
on how to write release notes. The bot will check your release notes
automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
* core 
  * Removed special cache for integers and used standard cache.
<!-- END RELEASE NOTES -->