@oscargus 2019-02-12
Looks good as far as I can tell. Good that you @NoumbissiValere based in on the earlier PR. As it involves a bit a git hassle, it is far too common that one just use the easier "copy-and-paste"-methodology.

What do you say @moorepants ? Is there something from #9294 still to be done to verify the (positive) impact? Seems like it would be nice to get this in if there are no obvious drawbacks.

Maybe one should try and run the example code from #15778 and see what it shows now? Would that be possible for you @NoumbissiValere ?