@oscargus 2019-02-12
The requested results from #15778 I just found in #15801 so that looks fine from a memory perspective.