@oscarbenjamin 2019-02-18
I just tried this PR and I was able to reproduce the objgraph results from #15801 so it looks like it solves the memory issue from #15778.

I also ran the whole non-slow test suite and got:
master: 1841 secs
this PR: 1665 secs
master with this PR merged in: 1811 secs

So it looks like this PR offers a 30 secs speedup over master but that might just be noise.

Also looks like there was a significant slowdown on master between the base of this branch and current tip...