@oscargus 2019-02-18
It may, at least to some extent, be that there are simply more tests now. Some non-slow tests were far from non-slow so if any of those are added since... But probably not the whole explanation.

Possible speed-up and fixed memory leak sounds good though.