@oscarbenjamin 2019-03-02
@NoumbissiValere sorry it has taken so long to merge this. I wasn't familiar with SymPy's caching code. I've taken a bit of a look at that and this and the related issues and this looks good to me.

I will merge this soon (just waiting until there isn't a build queue on Travis).